Add StorePath() to Resource interface
