+++ api/core/v2/adhoc.go
@@ -6,20 +6,20 @@ import (
 	"net/url"
 )
 
-// Validate returns an error if the name is not provided.
-func (a *AdhocRequest) Validate() error {
-	if a.Name == "" {
-		return errors.New("must provide check name")
-	}
-	return nil
-}
-
 // FixtureAdhocRequest returns a testing fixture for an AdhocRequest struct.
 func FixtureAdhocRequest(name string, subscriptions []string) *AdhocRequest {
 	return &AdhocRequest{
 		ObjectMeta:    NewObjectMeta(name, "default"),
 		Subscriptions: subscriptions,
 	}
+}
+
+// Validate returns an error if the name is not provided.
+func (a *AdhocRequest) Validate() error {
+	if a.Name == "" {
+		return errors.New("must provide check name")
+	}
+	return nil
 }
 
 // URIPath is the URI path component to the adhoc request.
