+++ api/core/v2/adhocpb_test.go
@@ -120,6 +120,14 @@ func TestAdhocRequestProtoCompactText(t *testing.T) {
 	}
 }
 
+func TestAdhocRequestFace(t *testing.T) {
+	popr := math_rand.New(math_rand.NewSource(time.Now().UnixNano()))
+	p := NewPopulatedAdhocRequest(popr, true)
+	msg := p.TestProto()
+	if !p.Equal(msg) {
+		t.Fatalf("%#v !Face Equal %#v", msg, p)
+	}
+}
 func TestAdhocRequestSize(t *testing.T) {
 	seed := time.Now().UnixNano()
 	popr := math_rand.New(math_rand.NewSource(seed))
