+++ api/core/v2/check_test.go
@@ -2,7 +2,6 @@ package v2
 
 import (
 	"encoding/json"
-	"sort"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -19,39 +18,6 @@ func TestCheckValidate(t *testing.T) {
 
 	c.Name = "test"
 
-	assert.NoError(t, c.Validate())
-}
-
-func TestCheckConfig(t *testing.T) {
-	var c CheckConfig
-
-	// Invalid name
-	assert.Error(t, c.Validate())
-	c.Name = "foo"
-
-	// Invalid interval
-	assert.Error(t, c.Validate())
-	c.Interval = 60
-
-	// Invalid command
-	assert.Error(t, c.Validate())
-	c.Command = "echo 'foo'"
-
-	// Invalid namespace
-	assert.Error(t, c.Validate())
-	c.Namespace = "default"
-
-	// Invalid ttl
-	c.Ttl = 10
-	assert.Error(t, c.Validate())
-
-	// Invalid output metric format
-	c.OutputMetricFormat = "foo"
-	assert.Error(t, c.Validate())
-	c.OutputMetricFormat = ""
-
-	// Valid check
-	c.Ttl = 90
 	assert.NoError(t, c.Validate())
 }
 
@@ -97,25 +63,6 @@ func TestMergeWith(t *testing.T) {
 	assert.Equal(t, newCheck.Status, newCheck.History[20].Status)
 }
 
-func TestProxyRequestsValidate(t *testing.T) {
-	var p ProxyRequests
-
-	// Invalid splay coverage
-	p.SplayCoverage = 150
-	assert.Error(t, p.Validate())
-	p.SplayCoverage = 0
-
-	// Invalid splay and splay coverage
-	p.Splay = true
-	assert.Error(t, p.Validate())
-	p.SplayCoverage = 90
-
-	p.EntityAttributes = []string{`entity.EntityClass == "proxy"`}
-
-	// Valid proxy request
-	assert.NoError(t, p.Validate())
-}
-
 func TestOutputMetricFormatValidate(t *testing.T) {
 	assert.NoError(t, ValidateOutputMetricFormat("nagios_perfdata"))
 	assert.NoError(t, ValidateOutputMetricFormat(NagiosOutputMetricFormat))
@@ -124,17 +71,6 @@ func TestOutputMetricFormatValidate(t *testing.T) {
 	assert.NoError(t, ValidateOutputMetricFormat(OpenTSDBOutputMetricFormat))
 	assert.Error(t, ValidateOutputMetricFormat("anything_else"))
 	assert.Error(t, ValidateOutputMetricFormat("NAGIOS_PERFDATA"))
-}
-
-func TestFixtureProxyRequests(t *testing.T) {
-	p := FixtureProxyRequests(true)
-
-	assert.Equal(t, true, p.Splay)
-	assert.Equal(t, uint32(90), p.SplayCoverage)
-	assert.NoError(t, p.Validate())
-
-	p.SplayCoverage = 0
-	assert.Error(t, p.Validate())
 }
 
 func TestCheckHasZeroIssuedMarshaled(t *testing.T) {
@@ -169,22 +105,6 @@ func TestCheckHasNonNilHandlers(t *testing.T) {
 	require.NotNil(t, c.Handlers)
 }
 
-func TestCheckConfigHasNonNilSubscriptions(t *testing.T) {
-	var c CheckConfig
-	b, err := json.Marshal(&c)
-	require.NoError(t, err)
-	require.NoError(t, json.Unmarshal(b, &c))
-	require.NotNil(t, c.Subscriptions)
-}
-
-func TestCheckConfigHasNonNilHandlers(t *testing.T) {
-	var c CheckConfig
-	b, err := json.Marshal(&c)
-	require.NoError(t, err)
-	require.NoError(t, json.Unmarshal(b, &c))
-	require.NotNil(t, c.Handlers)
-}
-
 func TestCheckFlapThresholdValidation(t *testing.T) {
 	c := FixtureCheck("foo")
 	// zero-valued flap threshold is valid
@@ -202,57 +122,4 @@ func TestCheckFlapThresholdValidation(t *testing.T) {
 	// low flap threshold > high flap threshold is invalid
 	c.LowFlapThreshold, c.HighFlapThreshold = 11, 10
 	assert.Error(t, c.Validate())
-}
-
-func TestCheckConfigFlapThresholdValidation(t *testing.T) {
-	c := FixtureCheckConfig("foo")
-	// zero-valued flap threshold is valid
-	c.LowFlapThreshold, c.HighFlapThreshold = 0, 0
-	assert.NoError(t, c.Validate())
-
-	// low flap threshold < high flap threshold is valid
-	c.LowFlapThreshold, c.HighFlapThreshold = 5, 10
-	assert.NoError(t, c.Validate())
-
-	// low flap threshold = high flap threshold is invalid
-	c.LowFlapThreshold, c.HighFlapThreshold = 10, 10
-	assert.Error(t, c.Validate())
-
-	// low flap threshold > high flap threshold is invalid
-	c.LowFlapThreshold, c.HighFlapThreshold = 11, 10
-	assert.Error(t, c.Validate())
-}
-
-func TestSortCheckConfigsByName(t *testing.T) {
-	a := FixtureCheckConfig("Abernathy")
-	b := FixtureCheckConfig("Bernard")
-	c := FixtureCheckConfig("Clementine")
-	d := FixtureCheckConfig("Dolores")
-
-	testCases := []struct {
-		name     string
-		inDir    bool
-		inChecks []*CheckConfig
-		expected []*CheckConfig
-	}{
-		{
-			name:     "Sorts ascending",
-			inDir:    true,
-			inChecks: []*CheckConfig{d, c, a, b},
-			expected: []*CheckConfig{a, b, c, d},
-		},
-		{
-			name:     "Sorts descending",
-			inDir:    false,
-			inChecks: []*CheckConfig{d, a, c, b},
-			expected: []*CheckConfig{d, c, b, a},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			sort.Sort(SortCheckConfigsByName(tc.inChecks, tc.inDir))
-			assert.EqualValues(t, tc.expected, tc.inChecks)
-		})
-	}
 }
