+++ api/core/v2/extension.go
@@ -2,9 +2,24 @@ package v2
 
 import (
 	"errors"
-	"fmt"
 	"net/url"
+	"path"
 )
+
+const (
+	// ExtensionsResource is the name of this resource type
+	ExtensionsResource = "extensions"
+)
+
+// StorePath returns the path prefix to extensions in the store
+func (e *Extension) StorePath() string {
+	return ExtensionsResource
+}
+
+// URIPath returns the path component of an extension URI.
+func (e *Extension) URIPath() string {
+	return path.Join(URLPrefix, "namespaces", url.PathEscape(e.Namespace), ExtensionsResource, url.PathEscape(e.Name))
+}
 
 // Validate validates the extension.
 func (e *Extension) Validate() error {
@@ -18,11 +33,6 @@ func (e *Extension) Validate() error {
 		return errors.New("empty namespace")
 	}
 	return nil
-}
-
-// URIPath returns the path component of an Extension URI.
-func (e *Extension) URIPath() string {
-	return fmt.Sprintf("/api/core/v2/namespaces/%s/extensions/%s", url.PathEscape(e.Namespace), url.PathEscape(e.Name))
 }
 
 // FixtureExtension given a name returns a valid extension for use in tests
