+++ api/core/v2/tessen.go
@@ -2,9 +2,11 @@ package v2
 
 import (
 	"fmt"
+	"path"
 )
 
 const (
+
 	// LowerBound is the minimum interval that tessen will phone home, in seconds
 	LowerBound = 60
 
@@ -14,9 +16,29 @@ const (
 	// DefaultTessenInterval is the default interval at which tessen will phone home, in seconds
 	DefaultTessenInterval = 1800
 
-	// TessenPath is the store and api path for tessen
-	TessenPath = "/tessen"
+	// TessenResource is the name of this resource type
+	TessenResource = "tessen"
 )
+
+// GetObjectMeta only exists here to fulfil the requirements of Resource
+func (t *TessenConfig) GetObjectMeta() ObjectMeta {
+	return ObjectMeta{}
+}
+
+// StorePath returns the path prefix to the Tessen config in the store
+func (t *TessenConfig) StorePath() string {
+	return TessenResource
+}
+
+// URIPath returns the path component of the Tessen config URI.
+func (t *TessenConfig) URIPath() string {
+	return path.Join(URLPrefix, TessenResource)
+}
+
+// Validate validates the TessenConfig.
+func (t *TessenConfig) Validate() error {
+	return nil
+}
 
 // ValidateInterval returns an error if the tessen interval is not within the upper and lower bound limits
 func ValidateInterval(freq uint32) error {
@@ -29,19 +51,4 @@ func ValidateInterval(freq uint32) error {
 // DefaultTessenConfig returns the default tessen configuration
 func DefaultTessenConfig() *TessenConfig {
 	return &TessenConfig{}
-}
-
-// URIPath returns the path component of a TessenConfig URI.
-func (t *TessenConfig) URIPath() string {
-	return fmt.Sprintf("/api/core/v2%s", TessenPath)
-}
-
-// Validate validates the TessenConfig.
-func (t *TessenConfig) Validate() error {
-	return nil
-}
-
-// GetObjectMeta only exists here to fulfil the requirements of Resource
-func (t *TessenConfig) GetObjectMeta() ObjectMeta {
-	return ObjectMeta{}
 }
