+++ api/core/v2/user.go
@@ -4,17 +4,29 @@ import (
 	"errors"
 	fmt "fmt"
 	"net/url"
+	"path"
 	"strconv"
 	"strings"
 )
 
-// FixtureUser returns a testing fixture for an Entity object.
-func FixtureUser(username string) *User {
-	return &User{
-		Username: username,
-		Password: "P@ssw0rd!",
-		Groups:   []string{"default"},
-	}
+const (
+	// UsersResource is the name of this resource type
+	UsersResource = "users"
+)
+
+// GetObjectMeta is a dummy implementation to meet the Resource interface.
+func (u *User) GetObjectMeta() ObjectMeta {
+	return ObjectMeta{}
+}
+
+// StorePath returns the path prefix to mutators in the store
+func (u *User) StorePath() string {
+	return UsersResource
+}
+
+// URIPath is the URI path component to a user.
+func (u *User) URIPath() string {
+	return path.Join(URLPrefix, UsersResource, url.PathEscape(u.Username))
 }
 
 // Validate returns an error if the entity is invalid.
@@ -39,14 +51,13 @@ func (u *User) ValidatePassword() error {
 	return nil
 }
 
-// URIPath is the URI path component to a user.
-func (u *User) URIPath() string {
-	return fmt.Sprintf("/api/core/v2/users/%s", url.PathEscape(u.Username))
-}
-
-// GetObjectMeta is a dummy implementation to meet the Resource interface.
-func (u *User) GetObjectMeta() ObjectMeta {
-	return ObjectMeta{}
+// FixtureUser returns a testing fixture for an Entity object.
+func FixtureUser(username string) *User {
+	return &User{
+		Username: username,
+		Password: "P@ssw0rd!",
+		Groups:   []string{"default"},
+	}
 }
 
 // UserFields returns a set of fields that represent that resource
